Component file reorganization
