+++ app/error.jsx
@@ -3,7 +3,7 @@
 import spanishdex from "@/public/logo.svg";
 import Image from "next/image";
 import Link from "next/link";
-import HomepageSection from "@/components/HomepageSection";
+import HomepageSection from "@/components/homepage/HomepageSection";
 import Button from "react-bootstrap/Button";
 import rightChevron from "@/public/icons/chevron_right.svg";
 import downChevron from "@/public/icons/chevron_down.svg";
